+++ NEWS.md
@@ -1,7 +1,7 @@
 # shinyHeatmap 0.2.2.9000
 
 ## Minor change
-- Adjust z-index for pushar overlay, pushbar and heatmap canvas.
+- Adjust z-index for pushbar and heatmap canvas.
 - Remove delay in R code.
 - More space between heatmap buttons (toggle and options).